apps/backend/requirements.txt
aiohttp==3.12.13
fastapi==0.115.14
ipython==9.3.0
langchain_chroma==0.2.4
langchain_community==0.3.27
langchain_core==0.3.68
langchain_openai==0.3.27
langchain_text_splitters==0.3.8
langgraph==0.5.1
pydantic==2.11.7
python-dotenv==1.1.1
